+++ CHANGELOG.md
@@ -7,3 +7,9 @@
  * **generate:module** - Add hook stub for hook_civicrm_caseTypes (in module.php.php and module.civix.php.php)
  * Add documentation links for hooks (using "@link")
  * Reformat civix source code based on CiviCRM's coding conventions
+
+### v14.??
+
+ * Removed options **civicrm_api3_conf_path**, **civicrm_api3_server**, **civicrm_api3_path**, **civicrm_api3_key**, and **civicrm_api3_api_key**
+ * **civix** will scan the directory tree to locate and bootstrap the CMS.
+ * **civix generate:module** - Automatically refresh extension list. Prompt for installation.